+++ i3configger/main.py
@@ -2,7 +2,7 @@ import logging
 import sys
 
 from i3configger import (
-    build, cli, daemonize, ipc, partials, paths, state, watch)
+    build, cli, daemonize, ipc, partials, paths, message, watch)
 
 log = logging.getLogger(__name__)
 
@@ -11,23 +11,22 @@ def main():
     args = cli.process_command_line()
     if args.kill:
         daemonize.exorcise()
-        sys.exit()
+        return 0
     configPath = paths.get_my_config_path(args.config)
     if args.message:
         p = paths.Paths(configPath)
-        state.process(p.state, partials.create(p.root), args.message)
+        prts = partials.create(p.root)
+        message.process(p.messages, prts, args.message)
     if daemonize.get_other_i3configgers():
         if not args.message:
-            sys.exit("i3configger already running. "
-                     "Did you want to send a message?")
-        log.info("watcher process is running - processing message")
-        sys.exit()
+            sys.exit("already running. Did you mean to send a message?")
+        return 0
     ipc.I3.set_msg_type(args.i3_refresh_msg)
     log.info("set i3 refresh method to %s", ipc.I3.refresh)
     ipc.Notify.set_notify_command(args.notify)
     if args.daemon:
         daemonize.daemonize(args.v, args.log, configPath)
-        sys.exit()
+        return 0
     if args.watch:
         try:
             watch.Watchman(configPath).watch()